requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.5
django-split-settings==1.2.0
Faker==19.11.0
flake8==6.1.0
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
